-Dialog variables + styles moved to teams theme
